enormous changes
This commit is the result of a campaign against numerical instabilities
I did a couple of things
1) I made sure that the surrogate posteriers and priors
available at the CLI have _identical_ support around zero to 32-bit
precision.
2) I made a new sampler for truncated normals which is a bit more stable
for my particular problem. It could still use some more work, but it is
a strong start.

There are many other little changes coming along for the ride on this
one. I may want to roll some of them back later, but they mostly amount
to extra, unused bits of code that I'm playing around with.